Skeleton pie: Move logic inside Skeleton class
