+++ packages/core/src/components/graphs/skeleton-pie.ts
@@ -1,9 +1,5 @@
 // MISC Imports
 import { Skeleton } from "./skeleton";
-import { DOMUtils } from "../../services";
-
-// D3 Imports
-import { arc } from "d3-shape";
 
 export class SkeletonPie extends Skeleton {
 	render() {
@@ -19,42 +15,9 @@ export class SkeletonPie extends Skeleton {
 	}
 
 	renderSkeleton() {
-		const outerRadius = this.computeRadius();
-		this.drawRing(outerRadius, 0);
+		const outerRadius = this.computeOuterRadius();
+		const innerRadius = 0;
+		this.drawRing(outerRadius, innerRadius);
 		this.setStyle("shimmer-areas");
-	}
-
-	drawRing(outerRadius: number, innerRadius: number) {
-		const svg = this.parent;
-		const { width, height } = DOMUtils.getSVGElementSize(svg, { useAttrs: true });
-		const container = DOMUtils.appendOrSelect(svg, "g.chart-skeleton");
-		const options = this.model.getOptions().pie;
-
-		const skeletonAreaContainer = DOMUtils.appendOrSelect(container, "rect.chart-skeleton-area-container")
-			.attr("width", width)
-			.attr("height", height)
-			.attr("fill", "none");
-
-		const arcPathGenerator = arc()
-			.innerRadius(innerRadius)
-			.outerRadius(outerRadius)
-			.startAngle(0)
-			.endAngle(Math.PI * 2);
-
-		// centering circle inside the container
-		const tcx = outerRadius + Math.abs(options.radiusOffset);
-		const tcy = outerRadius + (Math.min(width, height) - outerRadius * 2) / 2;
-
-		container.append("path")
-			.attr("class", "skeleton-area-shape")
-			.attr("transform", `translate(${tcx}, ${tcy})`)
-			.attr("d", arcPathGenerator);
-	}
-
-	computeRadius() {
-		const options = this.model.getOptions();
-		const { width, height } = DOMUtils.getSVGElementSize(this.parent, { useAttrs: true });
-		const radius = Math.min(width, height) / 2;
-		return radius + options.pie.radiusOffset;
 	}
 }
